requirements.txt
sentence-transformers
scikit-learn
autocorrect
transformers
google-genai